requirements.txt
graphql-py>=0.2.0
